Very Basic User Auth using Spring Security features
Implemented basic user auth. This means anytime a user makes an API call they must be authenticated (logged in). At this time the only issue with this code is using in-memory storage for users, which for my testing I had set to only user: "admin" and password: "pass". Need to figure out how to pull from db for this.