Updated index.js with cookie for user once they're singed in
Updated index.js with cookie for user once they're singed in. This will be used when we have favorites for a particular user. They can sign back in and have those same favorites displayed. This still needs to be implemented, which entails passing the cars that are favorited into an additional cookie tied to that user.